requirements.txt
# requirements.txt — Daftar dependency project

opencv-python>=4.5
Pillow>=9.0
numpy>=1.21
rembg>=2.0
pandas>=1.3
openpyxl>=3.0
